requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
faiss-cpu
sentence-transformers
